Update new data format - no duplet
Updated most code to new experminent form.
TODO:
- apply more info searching per curve
- rewrite the duplet and triplet filtering for the new data format (no
duplet filtering is present atm)

Issues:
NMRGlue is not reading the experiment correctly; causing a big loss in
resolution in the f2 dimension.
For database use; 4k points are used instead of the normal 16-32k. For
the sample; only 2k is used. It is abviously due to the shifting O1 that
NMRGlue (or Topspin) are having side effects.