Push mid-week 12
Mid week push with new features:
- the entire framework is working atm; after exams it will be changed
towards a class-like structure as Peter suggested.
- a database compile and load function are fully operational
- a function was added to quantify the similarity between two chunks;
this to implement a database search algorithm. This is using weight
functions but only for the sample! not yet for the database...
- the program is now fit to use with a single function to compare a
sample to the database
- multiple plotting functions have been implemented for quick testing
and debug (in the plotter.py)
- the function comparer function has multiple poss; however the usage of
the statistacal R-squared is proving to be highly efficient.
- The noise limit is now determined using a statistical analysis of the
noise-peaks hight and their standard deviation.

Problems with the current script:
- multiple alpha peaks for an unknown sample should be prefiltered; one
should be removed and the other one used. This is not implemented in the
current structure and should be implemented once we are using the
class-like structure.
- A punishment for unused functions of the database chunk should be
implemented; meaning if large peaks of a database chunk arent used; it
is not possible to assign this as a result. This can use the
pre-existing weight function used by the sample.
- theire should be an option for unasignable peaks of the sample. noise
peaks or small peaks might not get asigned as they are lacking in the
database. The script should now not to asign these and calculate a
punishment for this.
- The script is taking together certain alpha peaks from samples in case
there are multiple alpha peaks present. The script should keep them
seperated and try to asign a sugar to each alpha peak; to identify the
sample.

Extra functions:
- The database should go towards a dropbox-like location; so the script
can download the database from the internet in case that it got updated.
Otherwise it should continue to use the present database. This removes
the option for people to compile there own database; however we can
change it so that you can compile by "adding" or "recompile" to
overwrite current functions.

Chemistry side:
- All data seems to be correct for the database; however certain samples
have bad accuracy for the alpha region. This can be solved by using more
selective peaks; or trying to implement a 'multiple chunk analysis'
together. It might be poss. to add a filter that will see if other
chunks have the same alpha peak present; and filter this out. This
however is hard to implement and might not be feasable in the time frame
of the master thesis.

Presentation:
It has been discussed to allready give a presentation after the exams;
this to bring the readers allready a bit familiar with the subject; as
the IT department has very few chemistry knowledge; and the chemistry
department has no IT knowledge. It should start with an introduction
towards sugars in general and existing NMR techniques to identify the
sacharides.
For IT; we should explain the basic concepts of reading in the data;
performing the value analysis and the final chunk comparison. This will
be discused during the meeting of 15/12/2016.