Week -1: Classification
As was mentioned before; the code needed a classification.
This is a diffrent method in comparison to before; we no longer use
lists and dictionaries to store all the info; now we use four main
classes to store all the information:

--------------------------------|
Class I: The Experiment class |
--------------------------------|

This is used to read in the experiment and chunkify it. This class is
not stored in the database.
It contains all information on the experiment as different parameters.

--------------------------------|
Class II: The Chunk class        |
--------------------------------|

This class is sugar specific; in contradiction to the experiment class;
which is usually multiple sacharides.
This class is used to keep track of the data layer; which is not used
after integration.
The integration data is stored in the Values class.

--------------------------------|
Class III: The Values class       |
--------------------------------|

This is the main class; specific for each monosacharide. It does not
contain experiment specific information; however it contains the curves
which are typical for each sugar.
This is the class which we will store in the database; so that
experiments can be referenced against this.

--------------------------------|
Class IV: The Curve class        |
--------------------------------|

This is a specific curve of the Values; this will be tested upon
multiple fronts; to have more info.
Typical examples are the H1 value; the rico and so on.

A plotting function has been made for Class I till III; with each
different types of output. A comparator has allready been made for the
Curve class; yet it is still absent for the Values class.