Week IV - Framework
A framework was added for the new pulse program sequence; however; the
reading in of the data is still absent due to an actual Band Selective
3D TOCSY spectra.

The Band Selective 3D TOCSY's concept however; has allready been proven
by taking a Band Selective 2D TOCSY on sample JD #1. Here we can see
increased H1-seperation due to the second dimension. The measurement
time is significantly larger; therefor it is recommended to only use
this on complex saccharide mixtures where the H1 peaks of the
monosaccharides overlap.

Due to the extra work on the pulse program; no work will be done for now
on PyQT; the file will be merged into the main branch with no
repercussion.